fix/onboarding+dummy_enclosure
the generic enclosure partially handled some onboarding functions that dont belong there, in HolmesV the generic enclosure becomes essentially the GUI websocket service and no longer handles onboarding functionality

Fixes the readiness check to include pairing check, only report mycroft.ready after device is paired

HolmesV is meant to run in a modular fashion and might not have all services, this commit makes enclosure report ready after a timeout. A proper refactor is needed but breaks mycroft-core compatibility so it is most likely a WONTFIX. At least reduces the log spam... mycroft-core will wait forever and throw exceptions if not every service is loaded

The only possible issue here is a delayed mycroft.ready message, mycroft-core considers itself ready if all services are running and before pairing, HolmesV waits until they all load, but if they dont all load it sends the ready signal anyway after timeout, chatterbox handles this fine and its not critical :shrug:

authored-by: jarbasai <jarbasai@mailfence.com>
